requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-generativeai==0.3.2
python-dotenv==1.0.1
pydantic==2.6.1